[FIX] Comment 연관관계 저장 및 검증로직 임시수정

+++ src/main/java/com/sparta/springtrello/domain/comment/service/CommentService.java
@@ -9,9 +9,10 @@ import com.sparta.springtrello.domain.comment.dto.response.CommentSaveResponseDt
 import com.sparta.springtrello.domain.comment.entity.Comment;
 import com.sparta.springtrello.domain.comment.repository.CommentRepository;
 import com.sparta.springtrello.domain.member.entity.Member;
+import com.sparta.springtrello.domain.member.entity.MemberRole;
 import com.sparta.springtrello.domain.member.repository.MemberRepository;
 import com.sparta.springtrello.domain.user.dto.AuthUser;
-import com.sparta.springtrello.domain.user.repository.UserRepository;
+import com.sparta.springtrello.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,29 +23,26 @@ import org.springframework.transaction.annotation.Transactional;
 public class CommentService {
 
     private final CommentRepository commentRepository;
-    private final UserRepository userRepository;
     private final MemberRepository memberRepository;
+    private final UserService userService;
 
     public CommentSaveResponseDto saveComment(AuthUser authUser, Long cardId, CommentSaveRequestDto commentSaveRequestDto) {
 
         Long userId = authUser.getId();
+        userService.checkUser(userId);
 
-        if (userRepository.findById(userId).isPresent()) {
-            throw new IllegalArgumentException("유저 없음");
+        Card card = cardRepository.findById(cardId).orElseThrow(() ->
+                new IllegalArgumentException("card없음"));
+
+        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
+                new IllegalArgumentException("멤버 등록안됨"));
+
+
+        if (!member.getMemberRole().equals(MemberRole.CREATOR)) {
+            throw new IllegalArgumentException("CREATOR만 수정, 삭제할 수 있음.");
         }
 
-//        Card card = cardRepository.findById(cardId).orElseThrow(() ->
-//                new IllegalArgumentException("card없음"));
-//
-//        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
-//                new IllegalArgumentException("멤버 등록안됨"));
-//
-//        if (memberRepository.findRoleByUserId(userId).equals(Reader)) {
-//            throw new IllegalArgumentException("쓰기/수정 권한 없음");
-//        }
-
-        Comment comment = commentRepository.save(new Comment(commentSaveRequestDto));
-
+        Comment comment = commentRepository.save(new Comment(commentSaveRequestDto, card, member));
 
         return new CommentSaveResponseDto(comment);
     }
@@ -52,20 +50,17 @@ public class CommentService {
     public CommentEditResponseDto editComment(AuthUser authUser, Long id, CommentEditRequestDto commentEditRequestDto) {
 
         Long userId = authUser.getId();
-
-        if (userRepository.findById(userId).isPresent()) {
-            throw new IllegalArgumentException("유저 없음");
-        }
+        userService.checkUser(userId);
 
 //        Card card = cardRepository.findById(cardId).orElseThrow(() ->
 //                new IllegalArgumentException("card없음"));
-//
-//        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
-//                new IllegalArgumentException("멤버 등록안됨"));
-//
-//        if (memberRepository.findRoleByUserId(userId).equals(Reader)) {
-//            throw new IllegalArgumentException("쓰기/수정 권한 없음");
-//        }
+
+        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
+                new IllegalArgumentException("멤버 등록안됨"));
+
+        if (!member.getMemberRole().equals(MemberRole.CREATOR)) {
+            throw new IllegalArgumentException("CREATOR만 수정, 삭제할 수 있음.");
+        }
 
         Comment comment = commentRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("댓글 없음"));
@@ -78,19 +73,18 @@ public class CommentService {
 
     public void deleteComment(Long id, AuthUser authUser) {
 
-//        if (userRepository.findById(userId).isPresent()) {
-//            throw new IllegalArgumentException("유저 없음");
-//        }
-//
+        Long userId = authUser.getId();
+        userService.checkUser(userId);
+
 //        Card card = cardRepository.findById(cardId).orElseThrow(() ->
 //                new IllegalArgumentException("card없음"));
-//
-//        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
-//                new IllegalArgumentException("멤버 등록안됨"));
-//
-//        if (memberRepository.findRoleByUserId(userId).equals(Reader)) {
-//            throw new IllegalArgumentException("쓰기/수정 권한 없음");
-//        }
+
+        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
+                new IllegalArgumentException("멤버 등록안됨"));
+
+        if (!member.getMemberRole().equals(MemberRole.CREATOR)) {
+            throw new IllegalArgumentException("CREATOR만 수정, 삭제할 수 있음.");
+        }
 
         commentRepository.deleteById(id);
